+++ app/controllers/pedias_controller.rb
@@ -1,0 +1,8 @@
+class PediasController < ContentsController
+  expose(:model) { Pedia }
+  expose(:pedias) { current_group.pedias }
+  expose(:pedia)
+  expose(:contents) { pedias }
+  expose(:content) { pedia }
+
+end