+++ app/models/ability.rb
@@ -6,12 +6,14 @@ class Ability
     can :read, Page
     can :read, Chapter
     can :read, Scene
+    can :read, Character
+    can :read, Pedia
     cannot :index, Page
 
     if user.blank?
       can :create, User
-    elsif user.admin?
-      can :manage, :all
+    #elsif user.admin?
+    # can :manage, :all
     else
       # Regular user
       can :create, Seed
@@ -19,6 +21,8 @@ class Ability
       cannot :create, User
       can :manage, Page
       can :update, User, id: user.id
+      can :manage, Character
+      can :manage, Pedia
       can :manage, Chapter, user_id: user.id
       can :manage, Scene, user_id: user.id
       can :update, Scene