+++ app/models/seed.rb
@@ -6,4 +6,8 @@ class Seed < ActiveRecord::Base
 
   belongs_to :user
   belongs_to :group
+
+  def pages
+    []
+  end
 end